orchestrator: add ring buffer config, wire FIE filter policy env/flag, fix lint issues
- Add RingBufferSize to orchestrator config and expose via CLI flag and env var
- Wire RETINA_FIE_FILTER_POLICY to CLI and document it
- Update README with missing flags and env variables
- Suppress funlen warnings for run and agentHandler
- Fix RingBuffer constructor signature (int -> uint64 internally)

+++ internal/orchestrator/structures/ringbuffer.go
@@ -95,7 +95,7 @@ type RingBuffer[T any] struct {
 }
 
 // NewRingBuffer creates a new RingBuffer with the given capacity.
-func NewRingBuffer[T any](capacity uint64) (*RingBuffer[T], error) {
+func NewRingBuffer[T any](capacity int) (*RingBuffer[T], error) {
 	if capacity <= 0 {
 		return nil, fmt.Errorf("invalid argument: capacity must be greater than zero")
 	}
@@ -103,7 +103,7 @@ func NewRingBuffer[T any](capacity uint64) (*RingBuffer[T], error) {
 	return &RingBuffer[T]{
 		mutex:     mu,
 		buffer:    make([]*T, capacity),
-		capacity:  capacity,
+		capacity:  uint64(capacity),
 		consumers: make(map[*ringConsumer[T]]struct{}),
 		cond:      sync.NewCond(mu),
 	}, nil
